+++ src/api-execution/src/main/java/org/rbernalop/apiexecution/script/infrastructure/adapter/PythonRunner.java
@@ -1,8 +1,11 @@
 package org.rbernalop.apiexecution.script.infrastructure.adapter;
 
+import org.rbernalop.apiexecution.script.domain.exception.ExecutionException;
+import org.rbernalop.apiexecution.script.domain.exception.FileException;
 import org.rbernalop.apiexecution.script.domain.value_object.RunResult;
 import org.rbernalop.apiexecution.script.domain.port.ScriptRunner;
 import org.rbernalop.apiexecution.script.infrastructure.util.ShellRunner;
+import org.springframework.data.util.Pair;
 
 import java.io.File;
 import java.nio.file.Files;
@@ -17,12 +20,17 @@ public class PythonRunner implements ScriptRunner {
             file.createNewFile();
             Files.writeString(file.toPath(), code);
             String executionCommand = "python3 Main.py";
-            String executionResult = ShellRunner.executeCommand(executionCommand).getFirst();
+            Pair<String, Boolean> result = ShellRunner.executeCommand(executionCommand);
+            String executionResult = result.getFirst();
+            boolean executionSuccess = result.getSecond();
+            if(!executionSuccess) {
+                throw new ExecutionException(executionResult);
+            }
             runResult.setExecutionResult(executionResult);
 
             file.delete();
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            throw new FileException(e.getMessage());
         }
         return runResult;
     }
